+++ src/ui/messages/audio/layer-audio-message-model.js
@@ -163,7 +163,7 @@ export default class AudioModel extends MessageTypeModel {
   setupSlots() {
     const slots = [
       [
-        this.title || this.sourceUrl.replace(/(.*\/)?(.*?)(\..*)?$/, '$2') || this.getOneLineSummary(true),
+        this.title || this.sourceUrl.replace(/(.*\/)?(.*?)$/, '$2') || this.getOneLineSummary(true),
       ].filter(value => value),
       [
         this.artist, this.album, this.genre,
@@ -199,9 +199,10 @@ export default class AudioModel extends MessageTypeModel {
     let str = '';
 
     let hrs = Math.floor(this.duration / 3600);
-    if (!hrs) hrs = '00';
-    else if (hrs < 10) hrs = '0' + hrs;
-    str += hrs + ':';
+    if (hrs) {
+      if (hrs < 10) hrs = '0' + hrs;
+      str += hrs + ':';
+    }
 
     const afterHours = this.duration % 3600;
     let mins = Math.floor(afterHours / 60);